requirements.txt
pyautogui
mouse
keyboard
pyperclip
pillow
pynput